OUT2MSEED converter version 20170517 works on rev 1.16 OUT files
Rampant timing glitches in the original digitizer are now tracked and
fixed with this converter, which rolls whole directories into miniseed.
Make sure the vdaq.txt file is corrected beforehand. PNE2SAC is an older
2016 version that works. It relies on my old antiquated method of making
SAC files and then converting them to miniseed. It is for old digitized
PNE records that are resident in the office.